django_project/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2019.11.28
chardet==3.0.4
distlib==0.3.0
Django==3.0.5
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-localflavor==1.6.2
django-locations-base==0.3.4
djangorestframework==3.11.0
filelock==3.0.12
geographiclib==1.50
geopy==1.22.0
idna==2.9
importlib-metadata==1.5.0
Markdown==3.2.2
mysqlclient==1.4.6
numpy==1.18.1
Pillow==7.1.2
PyMySQL==0.9.3
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.5
zipp==3.0.0
